TestSystem/software/py/Magnus/requirements.txt
hypothesis>=3.82.1
dataset>=1.0.8
progressbar2>=2.38.0
timeout-decorator>=0.4.0
fire>=0.1.3
configparser>=3.5.0
colorama>=0.3.9
six>=1.11.0
numpy>= 1.10.0
pyusb==1.0.2
pyserial==3.4